Enforce layer protection on kernels folder
