+++ src/webviews/webview-side/interactive-common/redux/reducers/types.ts
@@ -3,7 +3,14 @@
 'use strict';
 
 import { IJupyterVariablesRequest } from '../../../../../kernels/variables/types';
-import { IShowDataViewer, InteractiveWindowMessages, IEditorContentChange } from '../../../../../platform/messageTypes';
+import {
+    IShowDataViewer,
+    InteractiveWindowMessages,
+    IEditorContentChange,
+    LoadIPyWidgetClassLoadAction,
+    ILoadIPyWidgetClassFailureAction,
+    NotifyIPyWidgetWidgetVersionNotSupportedAction
+} from '../../../../../platform/messageTypes';
 import { BaseReduxActionPayload } from '../../../../types';
 import { NativeKeyboardCommandTelemetry, NativeMouseCommandTelemetry } from '../../../common/constants';
 import { ActionWithPayload, ReducerArg } from '../../../react-common/reduxUtils';
@@ -226,17 +233,6 @@ export interface IOpenSettingsAction {
     setting: string | undefined;
 }
 
-export interface ILoadIPyWidgetClassFailureAction {
-    className: string;
-    moduleName: string;
-    moduleVersion: string;
-    cdnsUsed: boolean;
-    isOnline: boolean;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    error: any;
-    timedout: boolean;
-}
-
 export interface IVariableExplorerHeight {
     containerHeight: number;
     gridHeight: number;
@@ -249,16 +245,6 @@ export interface IVariableViewHeight {
 export type LoadIPyWidgetClassDisabledAction = {
     className: string;
     moduleName: string;
-    moduleVersion: string;
-};
-
-export type LoadIPyWidgetClassLoadAction = {
-    className: string;
-    moduleName: string;
-    moduleVersion: string;
-};
-export type NotifyIPyWidgetWidgetVersionNotSupportedAction = {
-    moduleName: 'qgrid';
     moduleVersion: string;
 };
